mrc-load-model-utils(extract-resolve-sliding-window-move,mechanical_provable): move resolve_sliding_window_size to load_model_utils module
